requirements.txt
tensorflow-cpu
opencv-python
numpy
matplotlib
albumentations
